+++ BrightID/src/utils/recovery.ts
@@ -1,7 +1,30 @@
+import ChannelAPI from '@/api/channelService';
+import { hash } from '@/utils/encoding';
 
-export const buildRecoveryChannelQrUrl = ({ aesKey, url, t }: RecoveryChannel) => {
+export const buildRecoveryChannelQrUrl = ({
+  aesKey,
+  url,
+  t,
+}: RecoveryChannel) => {
   const qrUrl = new URL(url.href);
   qrUrl.searchParams.append('aes', aesKey);
   qrUrl.searchParams.append('t', t);
   return qrUrl;
+};
+
+export const uploadRecoveryData = async (
+  recoveryData: RecoveryData,
+  channelApi: ChannelAPI,
+) => {
+  const channelId = hash(recoveryData.aesKey);
+  const dataObj = {
+    signingKey: recoveryData.publicKey,
+    timestamp: recoveryData.timestamp,
+  };
+  const data = JSON.stringify(dataObj);
+  await channelApi.upload({
+    channelId,
+    data,
+    dataId: 'data',
+  });
 };
